High Code Agent (proceed with caution)/pyproject.toml
[project]
name = "wesfarmers-slide-reviewer"
version = "0.1.0"
description = "Google ADK demo agent for reviewing Wesfarmers BD slide decks."
readme = "README.md"
authors = [
    { name = "Ajmal", email = "33590121+grandintegrator@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.24.1",
    "google-api-python-client>=2.189.0",
    "google-auth>=2.48.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "python-dotenv>=1.2.1",
]

[project.scripts]
wesfarmers-slide-reviewer = "wesfarmers_slide_reviewer:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["wesfarmers_slide_reviewer"]
